+++ jdbc/src/main/java/org/springframework/jdbc/datasource/DataSourceConfig.java
@@ -1,8 +1,7 @@
-package com.techcourse.config;
-
-import org.h2.jdbcx.JdbcDataSource;
+package org.springframework.jdbc.datasource;
 
 import java.util.Objects;
+import org.h2.jdbcx.JdbcDataSource;
 
 public class DataSourceConfig {
